// carpool-front/src/app/modules/openapi/model/absence-days.ts
/**
 * carpool-back API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { WeekDay } from './week-day';
import { WeekType } from './week-type';


export interface AbsenceDays { 
    id?: number;
    weekDay?: WeekDay;
    weekType?: WeekType;
}
export namespace AbsenceDays {
}